### 2021-03-04 #3087 @whedon
[PRE REVIEW]: IFermi: A python library for Fermi surface generation and analysis
**Submitting author:** @utf (<a href="http://orcid.org/0000-0002-4486-3321">Alex Ganose</a>)
**Repository:** <a href="https://github.com/fermisurfaces/IFermi" target ="_blank">https://github.com/fermisurfaces/IFermi</a>
**Version:** v.0.2.1
