+++ paper.bib
@@ -19,6 +19,7 @@
     volume = {68},
     number = {C},
     pages = {314--319},
+    doi = {10.1016/j.commatsci.2012.10.028},
 }
 
 @article{xcrysden,
@@ -29,7 +30,8 @@
     number={3-4},
     pages={176--179},
     year={1999},
-    publisher={Elsevier}
+    publisher={Elsevier},
+    doi = {10.1016/s1093-3263(99)00028-5},
 }
 
 @article{matplotlib,
@@ -40,7 +42,8 @@
     number={3},
     pages={90--95},
     year={2007},
-    publisher={IEEE}
+    publisher={IEEE},
+    doi={10.1109/mcse.2007.55}
 }
 
 @article{vasp,
