### 2021-03-04 #3087 @danielskatz
👋 @lucydot & @arosen93 - Would you be willing to review this submission for JOSS?
