### 2021-03-04 #3087 @arosen93
@danielskatz, I would be happy to review this submission for JOSS. For full disclosure, while I have significant experience with electronic structure calculations, I have personally not generated Fermi surfaces before (although have been interested in doing so!). In this regard, I would be providing a review from the perspective of a new user interested in generating/analyzing Fermi surfaces for the first time. If this is not preferred, don't hesitate to let me know.
