+++ paper.md
@@ -27,7 +27,7 @@ affiliations:
    index: 2
  - name: Molecular Foundry, Lawrence Berkeley National Laboratory, Berkeley, California 94720, USA
    index: 3
- - name: Department of Physics, University of Oxford, Oxford OX1 3PJ, UK
+ - name: Clarendon Laboratory, Department of Physics, University of Oxford, OX1 3PU, UK
    index: 4
 date: 1 March 2021
 bibliography: paper.bib
