### 2021-03-04 #3087 @lucydot
Hi @danielskatz I'm happy to review. I'm in the same situation as the other reviewer - I haven't generated Fermi surfaces before. Though I know enough textbook physics about Fermi surfaces that I think I should be able to give it an educated whirl.
