### 2021-03-04 c8a5c92
Update release workflow
