+++ .github/workflows/docs.yml
@@ -3,7 +3,7 @@ name: Build documentation
 on:
   push:
     branches:
-      - master
+      - main
 
 jobs:
   docs:
