+++ .github/workflows/tests.yml
@@ -3,11 +3,11 @@ name: Run tests
 on:
   push:
     branches:
-      - master
+      - main
 
   pull_request:
     branches:
-      - master
+      - main
 
 jobs:
   test:
