### 2021-03-05 246675c
Fix PR and bug templates
