+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,10 +1,7 @@
 ---
-name: Bug report
-about: Create a report to help us improve
-title: ''
-labels: ''
-assignees: ''
-
+name: Bug Report
+about: Create a bug report to help us improve IFermi
+title: "BUG:"
 ---
 
 **Describe the bug**
@@ -17,22 +14,11 @@ Steps to reproduce the behavior:
 3. Scroll down to '....'
 4. See error
 
+Provide any example files that are needed to reproduce the error,
+especially if the bug pertains to parsing of a file.
+
 **Expected behavior**
 A clear and concise description of what you expected to happen.
 
 **Screenshots**
 If applicable, add screenshots to help explain your problem.
-
-**Desktop (please complete the following information):**
- - OS: [e.g. iOS]
- - Browser [e.g. chrome, safari]
- - Version [e.g. 22]
-
-**Smartphone (please complete the following information):**
- - Device: [e.g. iPhone6]
- - OS: [e.g. iOS8.1]
- - Browser [e.g. stock browser, safari]
- - Version [e.g. 22]
-
-**Additional context**
-Add any other context about the problem here.
