+++ .github/ISSUE_TEMPLATE/config.yml
@@ -1,0 +1,5 @@
+blank_issues_enabled: false
+contact_links:
+  - name: IFermi Support Forum
+    url: https://matsci.org/c/ifermi
+    about: Please ask and answer questions here.
