+++ _sources/introduction.rst.txt
@@ -3,4 +3,4 @@ Introduction
 ============
 
 .. mdinclude:: ../../README.md
-   :start-line: 5
+   :start-line: 7
