### 2021-03-10 fermisurfaces/IFermi#17 @lucydot
JOSS Review: installation in a conda environment
RE: openjournals/joss-reviews#3089

I tried to install ifermi using my usual approach, which is within a new conda environment:

> conda create --name ifermi
conda activate ifermi
conda install pip
pip install ifermi

This did not work as I think the combination of conda/pip meant that there were problems with the `cmake` command needed to install the `boltztrap2` and `pyfftw` dependencies (full output attached).

So instead I did the following:

> conda create --name ifermi
conda activate ifermi
conda install -c conda-forge boltztrap2==20.7.1
conda install -c conda-forge pyfftw==0.12.0
conda install pip
pip install ifermi

where I've pinned the conda-forge installs to the versions specified in requirements. iFermi installed without errors, though I haven't been able to run the tests (will raise a separate issue).

This isn't a bug as I installed within a conda environment, which is not what is specified in the documentation. And the problem of conda/pip not playing nicely is one that goes well beyond this project. But as a decent proportion of researchers seem to be using conda I thought I'd flag it up. The optimum solution would be to distribute a conda package (though then you'd need to think about whether the dependencies are available on conda also...), but documenting the above (maybe on the message board) would help others who use conda.

[conda_install_output.txt](https://github.com/fermisurfaces/IFermi/files/6116728/conda_install_output.txt)
