### 2021-03-10 fermisurfaces/IFermi#18 @lucydot
JOSS Review: tests
RE: openjournals/joss-reviews#3089

Perhaps you could include information for people who want to run the unit tests locally? Might be useful for those where the standard "pip install" doesn't work smoothly.

[In the documentation](https://fermisurfaces.github.io/IFermi/contributing.html?highlight=test#how-to-make-a-great-pull-request) you state that there is CI. If it's publically accessible through Travis-CI or similar then perhaps you could provide a link so that users can verify the latest version is working as expected.
