### 2021-03-10 fermisurfaces/IFermi#19 @lucydot
JOSS review: Documentation
RE: openjournals/joss-reviews#3089

The documentation looks very thorough and useful. I noticed a couple of smaller things:

- There are no units for group velocity or area in the API documentation. It's in the CLI section, and is commented in the code, but for quick reference might be nice to give in the API reference also.
- update documentation [here](https://fermisurfaces.github.io/IFermi/contributing.html?highlight=test#reporting-bugs-getting-help-and-discussion) with a link to the help forum
- Consider highlighting papers where ifermi has already been used? From the JOSS paper I see that the iFermi authors have related publications in the last couple of years (Park 2020, Inzani 2021)
