### 2021-03-10 fermisurfaces/IFermi#17 @utf
Hi Lucy,

Thanks for this. I wonder if the problem is actually missing cmake? Strangely, on my system I can install using:

```
conda create --name ifermi python=3.8 pip
conda activate ifermi
pip install ifermi
```

But maybe a better approach would be to install numpy and cmake at the same time as pip:
```
conda create --name ifermi python=3.8 pip cmake numpy
conda activate ifermi
pip install ifermi
```

Would you be willing to try these to see if either works for you? I'd rather not pip to specific boltztrap2/pyfftw versions in the docs as that will be a bit of a pain to keep up-to-date with requirements.txt.
