### 2021-03-10 fermisurfaces/IFermi#18 @utf
Thanks for point this out. I've added a "running tests" section in the Readme. Unfortunately, you need to clone the GitHub repo to have access to the tests. I've pasted the process below:

First download the source and install the test requirements.
```
git clone https://github.com/fermisurfaces/IFermi.git
cd IFermi
pip install .[tests]
```
Next, the tests can be run using:
```
pytest
```
Would you be able to let me knew if that process works for you?
