### 2021-03-10 fermisurfaces/IFermi#19 @utf
Thanks again Lucy.

- I've added the units for area in the FermiSurface reference docs. For example [here](https://fermisurfaces.github.io/IFermi/ifermi.html#ifermi.surface.FermiSurface.area).
-
- I've added the units for group velocity in the [interpolation reference docs](https://fermisurfaces.github.io/IFermi/ifermi.html#ifermi.interpolate.FourierInterpolator.interpolate_bands) (i.e., when the group velocities are calculated). The rest of the code is agnostic to the actual projections (i.e., they could be spin, group velocity or something specified by the user, so I don't think it makes sense to add specific units in the general case).
- The contributing page has been updated to include a link to the support forum. Good catch!
- IFermi hasn't actually been used in a paper yet. The awkwardness of making Fermi surfaces in the papers you mentioned was actually a driving force for creating IFermi.

If this answers your questions would you be happy to close this issue?
