### 2021-03-10 fermisurfaces/IFermi#17 @ajsearle97
Hi Both. Thanks for bringing this up Lucy. I usually have to explicitly install cmake but not fftw. As in, I would run
'''
conda create --name ifermi pip cmake numpy
conda activate ifermi
pip install ifermi
'''
and it installs without any issues.
