### 2021-03-10 fermisurfaces/IFermi#17 @arosen93
Hi all! I tried @ajsearle97's approach on a Linux machine, and it worked successfully. In case you're curious, I tried it also on a Windows machine without success (whether I explicitly install fftw or not).

```
    Running setup.py install for BoltzTraP2 ... error
     command: 'C:\Users\asros\anaconda3\envs\ifermi\python.exe' -u -c 'import sys, setuptools, tokenize; sys.argv[0] = '"'"'C:\\Users\\asros\\AppData\\Local\\Temp\\pip-install-3fompnbq\\boltztrap2_fbd9d2312e2f499684af97e149f17ae9\\setup.py'"'"'; __file__='"'"'C:\\Users\\asros\\AppData\\Local\\Temp\\pip-install-3fompnbq\\boltztrap2_fbd9d2312e2f499684af97e149f17ae9\\setup.py'"'"';f=getattr(tokenize, '"'"'open'"'"', open)(__file__);code=f.read().replace('"'"'\r\n'"'"', '"'"'\n'"'"');f.close();exec(compile(code, __file__, '"'"'exec'"'"'))' install --record 'C:\Users\asros\AppData\Local\Temp\pip-record-8_x7vgej\install-record.txt' --single-version-externally-managed --compile --install-headers 'C:\Users\asros\anaconda3\envs\ifermi\Include\BoltzTraP2'
         cwd: C:\Users\asros\AppData\Local\Temp\pip-install-3fompnbq\boltztrap2_fbd9d2312e2f499684af97e149f17ae9\
    Complete output (46 lines):
    running install
    running build
    running build_py
    creating build
    creating build\lib.win-amd64-3.8
    creating build\lib.win-amd64-3.8\BoltzTraP2
    copying BoltzTraP2\bandlib.py -> build\lib.win-amd64-3.8\BoltzTraP2
    copying BoltzTraP2\dft.py -> build\lib.win-amd64-3.8\BoltzTraP2
    copying BoltzTraP2\fd.py -> build\lib.win-amd64-3.8\BoltzTraP2
    copying BoltzTraP2\fermisurface.py -> build\lib.win-amd64-3.8\BoltzTraP2
    copying BoltzTraP2\fite.py -> build\lib.win-amd64-3.8\BoltzTraP2
    copying BoltzTraP2\interface.py -> build\lib.win-amd64-3.8\BoltzTraP2
    copying BoltzTraP2\io.py -> build\lib.win-amd64-3.8\BoltzTraP2
    copying BoltzTraP2\misc.py -> build\lib.win-amd64-3.8\BoltzTraP2
    copying BoltzTraP2\serialization.py -> build\lib.win-amd64-3.8\BoltzTraP2
    copying BoltzTraP2\units.py -> build\lib.win-amd64-3.8\BoltzTraP2
    copying BoltzTraP2\version.py -> build\lib.win-amd64-3.8\BoltzTraP2
    copying BoltzTraP2\__init__.py -> build\lib.win-amd64-3.8\BoltzTraP2
    creating build\lib.win-amd64-3.8\BoltzTraP2\sphere
    copying BoltzTraP2\sphere\__init__.py -> build\lib.win-amd64-3.8\BoltzTraP2\sphere
    running build_ext
    running build_spglib
    About to create a new build directory for spglib
    About to run 'cmake' for spglib
    -- Building for: Visual Studio 16 2019
    CMake Deprecation Warning at CMakeLists.txt:1 (cmake_minimum_required):
      Compatibility with CMake < 2.8.12 will be removed from a future version of
      CMake.

      Update the VERSION argument <min> value or use a ...<max> suffix to tell

      CMake that the project does not need compatibility with older versions.

    -- Selecting Windows SDK version 10.0.18362.0 to target Windows 10.0.19042.
    -- The C compiler identification is MSVC 19.25.28612.0
    -- Detecting C compiler ABI info
    -- Detecting C compiler ABI info - done
    -- Check for working C compiler: C:/Program Files (x86)/Microsoft Visual Studio/2019/Community/VC/Tools/MSVC/14.25.28610/bin/Hostx64/x64/cl.exe - skipped
    -- Detecting C compile features
    -- Detecting C compile features - done
    -- Build type: Release
    -- Configuring done
    -- Generating done
    -- Build files have been written to: C:/Users/asros/AppData/Local/Temp/pip-install-3fompnbq/boltztrap2_fbd9d2312e2f499684af97e149f17ae9/external/spglib-1.9.9/build-g65vg68a
    About to run 'make' for spglib
    error: [WinError 2] The system cannot find the file specified
ERROR: Command errored out with exit status 1: 'C:\Users\asros\anaconda3\envs\ifermi\python.exe' -u -c 'import sys, setuptools, tokenize; sys.argv[0] = '"'"'C:\\Users\\asros\\AppData\\Local\\Temp\\pip-install-3fompnbq\\boltztrap2_fbd9d2312e2f499684af97e149f17ae9\\setup.py'"'"'; __file__='"'"'C:\\Users\\asros\\AppData\\Local\\Temp\\pip-install-3fompnbq\\boltztrap2_fbd9d2312e2f499684af97e149f17ae9\\setup.py'"'"';f=getattr(tokenize, '"'"'open'"'"', open)(__file__);code=f.read().replace('"'"'\r\n'"'"', '"'"'\n'"'"');f.close();exec(compile(code, __file__, '"'"'exec'"'"'))' install --record 'C:\Users\asros\AppData\Local\Temp\pip-record-8_x7vgej\install-record.txt' --single-version-externally-managed --compile --install-headers 'C:\Users\asros\anaconda3\envs\ifermi\Include\BoltzTraP2' Check the logs for full command output.
```
