### 2021-03-10 fermisurfaces/IFermi#17 @utf
Thanks @ajsearle97 and @arosen93. I've updated the README with Amy's instructions.
