### 2021-03-10 #3089 @arosen93
@utf: I just read through the paper. It is clearly written, has a nice introduction to the importance of Fermi surfaces, discusses prior software in this area, and makes it clear where IFermi fits within the existing software ecosystem. As such, I have no major suggestions for changes to the text.

I only have a couple very trivial comments on the paper, unrelated to the software or science underlying the work.
- In the Summary: "IFermi is Python library..." --> "IFermi is a Python library..."
- The "author" for the Plotly citation got a bit truncated (Plotly Technologies Inc. --> Inc., P.T.).
