+++ paper.bib
@@ -179,7 +179,7 @@
 }
 
 @online{plotly,
-    author = {Plotly Technologies Inc.},
+    author = {{Plotly Technologies Inc.}},
     title = {Collaborative data science},
     publisher = {Plotly Technologies Inc.},
     address = {Montreal, QC},
