### 2021-03-11 fermisurfaces/IFermi#17 @lucydot
Hi @utf I just followed instructions from @ajsearle97 and it worked perfectly on a Mac. I won't close the issue in case you want to follow up on the issue with the Windows install.
