### 2021-03-11 fermisurfaces/IFermi#17 @utf
Thanks Lucy. I will raise an issue on the BoltzTraP2 repo about windows installation.
