### 2021-03-12 fermisurfaces/IFermi#17 @arosen93
Just a head's up: I have had trouble with the installation on a few Linux machines, each with different issues regarding the dependencies. This, again, doesn't have to do with IFermi directly so I'm not suggesting anything be done about this for the JOSS review, but you should be aware that some users might report issues.

On one machine, I received the following issue tracing back to pyfftw:
```
   command: /home/asr731/software/anaconda/envs/ifermi/bin/python3.9 -u -c 'import sys, setuptools, tokenize; sys.argv[0] = '"'"'/tmp/pip-install-5pz058_p/pyfftw_fde140813af74f09a37a9b00629bb001/setup.py'"'"'; __file__='"'"'/tmp/pip-install-5pz058_p/pyfftw_fde140813af74f09a37a9b00629bb001/setup.py'"'"';f=getattr(tokenize, '"'"'open'"'"', open)(__file__);code=f.read().replace('"'"'\r\n'"'"', '"'"'\n'"'"');f.close();exec(compile(code, __file__, '"'"'exec'"'"'))' bdist_wheel -d /tmp/pip-wheel-5zlk7gg1
       cwd: /tmp/pip-install-5pz058_p/pyfftw_fde140813af74f09a37a9b00629bb001/
  Complete output (30 lines):
  running bdist_wheel
  running build
  running build_py
  creating build
  creating build/lib.linux-x86_64-3.9
  creating build/lib.linux-x86_64-3.9/pyfftw
  copying pyfftw/__init__.py -> build/lib.linux-x86_64-3.9/pyfftw
  copying pyfftw/_version.py -> build/lib.linux-x86_64-3.9/pyfftw
  copying pyfftw/config.py -> build/lib.linux-x86_64-3.9/pyfftw
  creating build/lib.linux-x86_64-3.9/pyfftw/builders
  copying pyfftw/builders/__init__.py -> build/lib.linux-x86_64-3.9/pyfftw/builders
  copying pyfftw/builders/_utils.py -> build/lib.linux-x86_64-3.9/pyfftw/builders
  copying pyfftw/builders/builders.py -> build/lib.linux-x86_64-3.9/pyfftw/builders
  creating build/lib.linux-x86_64-3.9/pyfftw/interfaces
  copying pyfftw/interfaces/__init__.py -> build/lib.linux-x86_64-3.9/pyfftw/interfaces
  copying pyfftw/interfaces/_utils.py -> build/lib.linux-x86_64-3.9/pyfftw/interfaces
  copying pyfftw/interfaces/cache.py -> build/lib.linux-x86_64-3.9/pyfftw/interfaces
  copying pyfftw/interfaces/dask_fft.py -> build/lib.linux-x86_64-3.9/pyfftw/interfaces
  copying pyfftw/interfaces/numpy_fft.py -> build/lib.linux-x86_64-3.9/pyfftw/interfaces
  copying pyfftw/interfaces/scipy_fft.py -> build/lib.linux-x86_64-3.9/pyfftw/interfaces
  copying pyfftw/interfaces/scipy_fftpack.py -> build/lib.linux-x86_64-3.9/pyfftw/interfaces
  UPDATING build/lib.linux-x86_64-3.9/pyfftw/_version.py
  set build/lib.linux-x86_64-3.9/pyfftw/_version.py to '0.12.0'
  running build_ext
  Build pyFFTW with support for FFTW with
  double precision + openMP
  single precision + openMP
  long precision + openMP
  building 'pyfftw.pyfftw' extension
  error: unknown file type '.pyx' (from 'pyfftw/pyfftw.pyx')
  ERROR: Failed building wheel for pyfftw
  Running setup.py clean for pyfftw
Successfully built pymatgen retrying BoltzTraP2 future meshcut numpy-stl
Failed to build pyfftw
Installing collected packages: six, python-dateutil, pyparsing, pillow, kiwisolver, cycler, urllib3, scipy, ruamel.yaml.clib, retrying, pytz, python-utils, mpmath, matplotlib, idna, future, decorator, chardet, cftime, uncertainties, tifffile, tabulate, sympy, spglib, ruamel.yaml, requests, PyWavelets, plotly, pandas, palettable, numpy-stl, networkx, netCDF4, monty, imageio, ase, trimesh, scikit-image, pymatgen, pyfftw, psutil, meshcut, click, BoltzTraP2, ifermi
    Running setup.py install for pyfftw ... error
     command: /home/asr731/software/anaconda/envs/ifermi/bin/python3.9 -u -c 'import sys, setuptools, tokenize; sys.argv[0] = '"'"'/tmp/pip-install-5pz058_p/pyfftw_fde140813af74f09a37a9b00629bb001/setup.py'"'"'; __file__='"'"'/tmp/pip-install-5pz058_p/pyfftw_fde140813af74f09a37a9b00629bb001/setup.py'"'"';f=getattr(tokenize, '"'"'open'"'"', open)(__file__);code=f.read().replace('"'"'\r\n'"'"', '"'"'\n'"'"');f.close();exec(compile(code, __file__, '"'"'exec'"'"'))' install --record /tmp/pip-record-8aszlkbo/install-record.txt --single-version-externally-managed --compile --install-headers /home/asr731/software/anaconda/envs/ifermi/include/python3.9/pyfftw
         cwd: /tmp/pip-install-5pz058_p/pyfftw_fde140813af74f09a37a9b00629bb001/
    Complete output (12 lines):
    running install
    running build
    running build_py
    UPDATING build/lib.linux-x86_64-3.9/pyfftw/_version.py
    set build/lib.linux-x86_64-3.9/pyfftw/_version.py to '0.12.0'
    running build_ext
    Build pyFFTW with support for FFTW with
    double precision + openMP
    single precision + openMP
    long precision + openMP
    building 'pyfftw.pyfftw' extension
    error: unknown file type '.pyx' (from 'pyfftw/pyfftw.pyx')
ERROR: Command errored out with exit status 1: /home/asr731/software/anaconda/envs/ifermi/bin/python3.9 -u -c 'import sys, setuptools, tokenize; sys.argv[0] = '"'"'/tmp/pip-install-5pz058_p/pyfftw_fde140813af74f09a37a9b00629bb001/setup.py'"'"'; __file__='"'"'/tmp/pip-install-5pz058_p/pyfftw_fde140813af74f09a37a9b00629bb001/setup.py'"'"';f=getattr(tokenize, '"'"'open'"'"', open)(__file__);code=f.read().replace('"'"'\r\n'"'"', '"'"'\n'"'"');f.close();exec(compile(code, __file__, '"'"'exec'"'"'))' install --record /tmp/pip-record-8aszlkbo/install-record.txt --single-version-externally-managed --compile --install-headers /home/asr731/software/anaconda/envs/ifermi/include/python3.9/pyfftw Check the logs for full command output.
```

On a second machine, I received the following issue tracing back to Pymatgen:
```
...
creating build/temp.linux-x86_64-3.9/pymatgen/optimization
  gcc -pthread -B /home/rosen/software/anaconda3/envs/ifermi/compiler_compat -Wl,--sysroot=/ -Wno-unused-result -Wsign-compare -DNDEBUG -O2 -Wall -fPIC -O2 -isystem /home/rosen/software/anaconda3/envs/ifermi/include -fno-semantic-interposition -fPIC -O2 -isystem /home/rosen/software/anaconda3/envs/ifermi/include -fno-semantic-interposition -fPIC -I/home/rosen/software/anaconda3/envs/ifermi/include/python3.9 -I/tmp/pip-build-env-0kztei28/overlay/lib/python3.9/site-packages/numpy/core/include -c pymatgen/optimization/linear_assignment.c -o build/temp.linux-x86_64-3.9/pymatgen/optimization/linear_assignment.o
  gcc: error: unrecognized command line option ‘-fno-semantic-interposition’
  gcc: error: unrecognized command line option ‘-fno-semantic-interposition’
  error: command '/usr/bin/gcc' failed with exit code 1
  ERROR: Failed building wheel for pymatgen
   command: /home/rosen/software/anaconda3/envs/ifermi/bin/python -u -c 'import sys, setuptools, tokenize; sys.argv[0] = '"'"'/tmp/pip-install-yu21mx8_/boltztrap2_de6205c45a9045aaa49714aeb2a0aa06/setup.py'"'"'; __file__='"'"'/tmp/pip-install-yu21mx8_/boltztrap2_de6205c45a9045aaa49714aeb2a0aa06/setup.py'"'"';f=getattr(tokenize, '"'"'open'"'"', open)(__file__);code=f.read().replace('"'"'\r\n'"'"', '"'"'\n'"'"');f.close();exec(compile(code, __file__, '"'"'exec'"'"'))' bdist_wheel -d /tmp/pip-wheel-i7pyyztf
       cwd: /tmp/pip-install-yu21mx8_/boltztrap2_de6205c45a9045aaa49714aeb2a0aa06/
  Complete output (122 lines):
  running bdist_wheel
  running build
  running build_py
  creating build
  creating build/lib.linux-x86_64-3.9
  creating build/lib.linux-x86_64-3.9/BoltzTraP2
  copying BoltzTraP2/__init__.py -> build/lib.linux-x86_64-3.9/BoltzTraP2
  copying BoltzTraP2/bandlib.py -> build/lib.linux-x86_64-3.9/BoltzTraP2
  copying BoltzTraP2/dft.py -> build/lib.linux-x86_64-3.9/BoltzTraP2
  copying BoltzTraP2/fd.py -> build/lib.linux-x86_64-3.9/BoltzTraP2
  copying BoltzTraP2/fermisurface.py -> build/lib.linux-x86_64-3.9/BoltzTraP2
  copying BoltzTraP2/fite.py -> build/lib.linux-x86_64-3.9/BoltzTraP2
  copying BoltzTraP2/interface.py -> build/lib.linux-x86_64-3.9/BoltzTraP2
  copying BoltzTraP2/io.py -> build/lib.linux-x86_64-3.9/BoltzTraP2
  copying BoltzTraP2/misc.py -> build/lib.linux-x86_64-3.9/BoltzTraP2
  copying BoltzTraP2/serialization.py -> build/lib.linux-x86_64-3.9/BoltzTraP2
  copying BoltzTraP2/units.py -> build/lib.linux-x86_64-3.9/BoltzTraP2
  copying BoltzTraP2/version.py -> build/lib.linux-x86_64-3.9/BoltzTraP2
  creating build/lib.linux-x86_64-3.9/BoltzTraP2/sphere
  copying BoltzTraP2/sphere/__init__.py -> build/lib.linux-x86_64-3.9/BoltzTraP2/sphere
  running build_ext
  running build_spglib
  About to create a new build directory for spglib
  About to run 'cmake' for spglib
  CMake Deprecation Warning at CMakeLists.txt:1 (cmake_minimum_required):
    Compatibility with CMake < 2.8.12 will be removed from a future version of
    CMake.

    Update the VERSION argument <min> value or use a ...<max> suffix to tell
    CMake that the project does not need compatibility with older versions.

  -- The C compiler identification is GNU 4.8.5
  -- Detecting C compiler ABI info
  -- Detecting C compiler ABI info - done
  -- Check for working C compiler: /usr/bin/cc - skipped
  -- Detecting C compile features
  -- Detecting C compile features - done
  -- Build type: Release
  -- Configuring done
  -- Generating done
  -- Build files have been written to: /tmp/pip-install-yu21mx8_/boltztrap2_de6205c45a9045aaa49714aeb2a0aa06/external/spglib-1.9.9/build-uugrgrdg
  About to run 'make' for spglib
  Scanning dependencies of target symspg
  [  2%] Building C object CMakeFiles/symspg.dir/src/arithmetic.o
  [  5%] Building C object CMakeFiles/symspg.dir/src/cell.o
  [  7%] Building C object CMakeFiles/symspg.dir/src/debug.o
  [ 10%] Building C object CMakeFiles/symspg.dir/src/delaunay.o
  [ 12%] Building C object CMakeFiles/symspg.dir/src/hall_symbol.o
  [ 15%] Building C object CMakeFiles/symspg.dir/src/kgrid.o
  [ 17%] Building C object CMakeFiles/symspg.dir/src/kpoint.o
  /tmp/pip-install-yu21mx8_/boltztrap2_de6205c45a9045aaa49714aeb2a0aa06/external/spglib-1.9.9/src/kpoint.c: In function ‘get_ir_reciprocal_mesh_normal’:
  /tmp/pip-install-yu21mx8_/boltztrap2_de6205c45a9045aaa49714aeb2a0aa06/external/spglib-1.9.9/src/kpoint.c:536:0: warning: ignoring #pragma omp parallel [-Wunknown-pragmas]
   #pragma omp parallel for private(j, grid_point_rot, address_double, address_double_rot)
   ^
  /tmp/pip-install-yu21mx8_/boltztrap2_de6205c45a9045aaa49714aeb2a0aa06/external/spglib-1.9.9/src/kpoint.c: In function ‘get_ir_reciprocal_mesh_distortion’:
  /tmp/pip-install-yu21mx8_/boltztrap2_de6205c45a9045aaa49714aeb2a0aa06/external/spglib-1.9.9/src/kpoint.c:577:0: warning: ignoring #pragma omp parallel [-Wunknown-pragmas]
   #pragma omp parallel for private(j, k, grid_point_rot, address_double, address_double_rot)
   ^
  /tmp/pip-install-yu21mx8_/boltztrap2_de6205c45a9045aaa49714aeb2a0aa06/external/spglib-1.9.9/src/kpoint.c: In function ‘get_num_ir’:
  /tmp/pip-install-yu21mx8_/boltztrap2_de6205c45a9045aaa49714aeb2a0aa06/external/spglib-1.9.9/src/kpoint.c:618:0: warning: ignoring #pragma omp parallel [-Wunknown-pragmas]
   #pragma omp parallel for reduction(+:num_ir)
   ^
  [ 20%] Building C object CMakeFiles/symspg.dir/src/mathfunc.o
  [ 22%] Building C object CMakeFiles/symspg.dir/src/niggli.o
  [ 25%] Building C object CMakeFiles/symspg.dir/src/pointgroup.o
  [ 27%] Building C object CMakeFiles/symspg.dir/src/primitive.o
  [ 30%] Building C object CMakeFiles/symspg.dir/src/refinement.o
  [ 32%] Building C object CMakeFiles/symspg.dir/src/site_symmetry.o
  [ 35%] Building C object CMakeFiles/symspg.dir/src/sitesym_database.o
  [ 37%] Building C object CMakeFiles/symspg.dir/src/spacegroup.o
  [ 40%] Building C object CMakeFiles/symspg.dir/src/spg_database.o
  [ 42%] Building C object CMakeFiles/symspg.dir/src/spglib.o
  [ 45%] Building C object CMakeFiles/symspg.dir/src/spin.o
  [ 47%] Building C object CMakeFiles/symspg.dir/src/symmetry.o
  [ 50%] Linking C shared library libsymspg.so
  [ 50%] Built target symspg
  Scanning dependencies of target symspg_static
  [ 52%] Building C object CMakeFiles/symspg_static.dir/src/arithmetic.o
  [ 55%] Building C object CMakeFiles/symspg_static.dir/src/cell.o
  [ 57%] Building C object CMakeFiles/symspg_static.dir/src/debug.o
  [ 60%] Building C object CMakeFiles/symspg_static.dir/src/delaunay.o
  [ 62%] Building C object CMakeFiles/symspg_static.dir/src/hall_symbol.o
  [ 65%] Building C object CMakeFiles/symspg_static.dir/src/kgrid.o
  [ 67%] Building C object CMakeFiles/symspg_static.dir/src/kpoint.o
  /tmp/pip-install-yu21mx8_/boltztrap2_de6205c45a9045aaa49714aeb2a0aa06/external/spglib-1.9.9/src/kpoint.c: In function ‘get_ir_reciprocal_mesh_normal’:
  /tmp/pip-install-yu21mx8_/boltztrap2_de6205c45a9045aaa49714aeb2a0aa06/external/spglib-1.9.9/src/kpoint.c:536:0: warning: ignoring #pragma omp parallel [-Wunknown-pragmas]
   #pragma omp parallel for private(j, grid_point_rot, address_double, address_double_rot)
   ^
  /tmp/pip-install-yu21mx8_/boltztrap2_de6205c45a9045aaa49714aeb2a0aa06/external/spglib-1.9.9/src/kpoint.c: In function ‘get_ir_reciprocal_mesh_distortion’:
  /tmp/pip-install-yu21mx8_/boltztrap2_de6205c45a9045aaa49714aeb2a0aa06/external/spglib-1.9.9/src/kpoint.c:577:0: warning: ignoring #pragma omp parallel [-Wunknown-pragmas]
   #pragma omp parallel for private(j, k, grid_point_rot, address_double, address_double_rot)
   ^
  /tmp/pip-install-yu21mx8_/boltztrap2_de6205c45a9045aaa49714aeb2a0aa06/external/spglib-1.9.9/src/kpoint.c: In function ‘get_num_ir’:
  /tmp/pip-install-yu21mx8_/boltztrap2_de6205c45a9045aaa49714aeb2a0aa06/external/spglib-1.9.9/src/kpoint.c:618:0: warning: ignoring #pragma omp parallel [-Wunknown-pragmas]
   #pragma omp parallel for reduction(+:num_ir)
   ^
  [ 70%] Building C object CMakeFiles/symspg_static.dir/src/mathfunc.o
  [ 72%] Building C object CMakeFiles/symspg_static.dir/src/niggli.o
  [ 75%] Building C object CMakeFiles/symspg_static.dir/src/pointgroup.o
  [ 77%] Building C object CMakeFiles/symspg_static.dir/src/primitive.o
  [ 80%] Building C object CMakeFiles/symspg_static.dir/src/refinement.o
  [ 82%] Building C object CMakeFiles/symspg_static.dir/src/site_symmetry.o
  [ 85%] Building C object CMakeFiles/symspg_static.dir/src/sitesym_database.o
  [ 87%] Building C object CMakeFiles/symspg_static.dir/src/spacegroup.o
  [ 90%] Building C object CMakeFiles/symspg_static.dir/src/spg_database.o
  [ 92%] Building C object CMakeFiles/symspg_static.dir/src/spglib.o
  [ 95%] Building C object CMakeFiles/symspg_static.dir/src/spin.o
  [ 97%] Building C object CMakeFiles/symspg_static.dir/src/symmetry.o
  [100%] Linking C static library libsymspg.a
  [100%] Built target symspg_static
  copying libsymspg.a -> /tmp/pip-install-yu21mx8_/boltztrap2_de6205c45a9045aaa49714aeb2a0aa06/external/spglib-1.9.9
  About to remove the spglib build directory
  removing '/tmp/pip-install-yu21mx8_/boltztrap2_de6205c45a9045aaa49714aeb2a0aa06/external/spglib-1.9.9/build-uugrgrdg' (and everything under it)
  building 'BoltzTraP2.sphere.frontend' extension
  creating build/temp.linux-x86_64-3.9
  creating build/temp.linux-x86_64-3.9/BoltzTraP2
  creating build/temp.linux-x86_64-3.9/BoltzTraP2/sphere
  gcc -pthread -B /home/rosen/software/anaconda3/envs/ifermi/compiler_compat -Wl,--sysroot=/ -Wno-unused-result -Wsign-compare -DNDEBUG -O2 -Wall -fPIC -O2 -isystem /home/rosen/software/anaconda3/envs/ifermi/include -fno-semantic-interposition -fPIC -O2 -isystem /home/rosen/software/anaconda3/envs/ifermi/include -fno-semantic-interposition -fPIC -I/home/rosen/software/anaconda3/envs/ifermi/lib/python3.9/site-packages/numpy/core/include -I/tmp/pip-install-yu21mx8_/boltztrap2_de6205c45a9045aaa49714aeb2a0aa06/external/spglib-1.9.9/src -I/tmp/pip-install-yu21mx8_/boltztrap2_de6205c45a9045aaa49714aeb2a0aa06/external/eigen-eigen-3215c06819b9 -I/home/rosen/software/anaconda3/envs/ifermi/include/python3.9 -c BoltzTraP2/sphere/backend.cpp -o build/temp.linux-x86_64-3.9/BoltzTraP2/sphere/backend.o -std=c++11
  gcc: error: unrecognized command line option ‘-fno-semantic-interposition’
  gcc: error: unrecognized command line option ‘-fno-semantic-interposition’
  error: command '/usr/bin/gcc' failed with exit code 1
  ERROR: Failed building wheel for BoltzTraP2
   command: /home/rosen/software/anaconda3/envs/ifermi/bin/python -u -c 'import sys, setuptools, tokenize; sys.argv[0] = '"'"'/tmp/pip-install-yu21mx8_/pyfftw_f13a7cd7d0064791b63a7b9f6485e500/setup.py'"'"'; __file__='"'"'/tmp/pip-install-yu21mx8_/pyfftw_f13a7cd7d0064791b63a7b9f6485e500/setup.py'"'"';f=getattr(tokenize, '"'"'open'"'"', open)(__file__);code=f.read().replace('"'"'\r\n'"'"', '"'"'\n'"'"');f.close();exec(compile(code, __file__, '"'"'exec'"'"'))' bdist_wheel -d /tmp/pip-wheel-ry1r5rjb
       cwd: /tmp/pip-install-yu21mx8_/pyfftw_f13a7cd7d0064791b63a7b9f6485e500/
  Complete output (25 lines):
  running bdist_wheel
  running build
  running build_py
  creating build
  creating build/lib.linux-x86_64-3.9
  creating build/lib.linux-x86_64-3.9/pyfftw
  copying pyfftw/__init__.py -> build/lib.linux-x86_64-3.9/pyfftw
  copying pyfftw/_version.py -> build/lib.linux-x86_64-3.9/pyfftw
  copying pyfftw/config.py -> build/lib.linux-x86_64-3.9/pyfftw
  creating build/lib.linux-x86_64-3.9/pyfftw/builders
  copying pyfftw/builders/__init__.py -> build/lib.linux-x86_64-3.9/pyfftw/builders
  copying pyfftw/builders/_utils.py -> build/lib.linux-x86_64-3.9/pyfftw/builders
  copying pyfftw/builders/builders.py -> build/lib.linux-x86_64-3.9/pyfftw/builders
  creating build/lib.linux-x86_64-3.9/pyfftw/interfaces
  copying pyfftw/interfaces/__init__.py -> build/lib.linux-x86_64-3.9/pyfftw/interfaces
  copying pyfftw/interfaces/_utils.py -> build/lib.linux-x86_64-3.9/pyfftw/interfaces
  copying pyfftw/interfaces/cache.py -> build/lib.linux-x86_64-3.9/pyfftw/interfaces
  copying pyfftw/interfaces/dask_fft.py -> build/lib.linux-x86_64-3.9/pyfftw/interfaces
  copying pyfftw/interfaces/numpy_fft.py -> build/lib.linux-x86_64-3.9/pyfftw/interfaces
  copying pyfftw/interfaces/scipy_fft.py -> build/lib.linux-x86_64-3.9/pyfftw/interfaces
  copying pyfftw/interfaces/scipy_fftpack.py -> build/lib.linux-x86_64-3.9/pyfftw/interfaces
  UPDATING build/lib.linux-x86_64-3.9/pyfftw/_version.py
  set build/lib.linux-x86_64-3.9/pyfftw/_version.py to '0.12.0'
  running build_ext
  error: Could not find the FFTW header 'fftw3.h'
  ERROR: Failed building wheel for pyfftw
ERROR: Could not build wheels for pymatgen which use PEP 517 and cannot be installed directly
```
