### 2021-03-12 fermisurfaces/IFermi#20 @arosen93
JOSS Review: Plotting
RE: openjournals/joss-reviews#3089

**Describe the bug**
When I use `ifermi plot`, I receive the following error message:
```

Start : This command cannot be run due to the error: The system cannot find the file specified.
At line:1 char:1
+ Start "file:///home/arosen/ifermi/fermi-surface.html"
+ ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
    + CategoryInfo          : InvalidOperation: (:) [Start-Process], InvalidOperationException
    + FullyQualifiedErrorId : InvalidOperationException,Microsoft.PowerShell.Commands.StartProcessCommand
```

**To Reproduce**
1. Install IFermi
2. Run `ifermi plot` on an example `vasprun.xml` (e.g. [vasprun.zip](https://github.com/fermisurfaces/IFermi/files/6133603/vasprun.zip)).

**Expected behavior**
A plot. There is no plot.

**Hypothesis**
Again, perhaps this is a Windows thing. I am running IFermi on the Ubuntu subsystem for Windows since two of my common Linux workstation had issues with the installation (see https://github.com/fermisurfaces/IFermi/issues/17).
