### 2021-03-12 fermisurfaces/IFermi#17 @utf
Thanks for the report @arosen93. Would you be willing to try the following installation process on your machines to see if this fixes the issues:

```
conda create --name ifermi pip cmake numpy fftw
conda activate ifermi
conda install -c hcc boltztrap2
conda install -c conda-forge pymatgen
pip install ifermi
```

Worst comes to worst we can release a conda package as @lucydot suggested.
