### 2021-03-12 fermisurfaces/IFermi#20 @utf
Hi @arosen93, that is a very strange error message. It doesn't look like something from python. Is there anything else printed from that command? Lastly, can you confirm that no files have been written to: /home/arosen/ifermi/ ?
