### 2021-03-12 fermisurfaces/IFermi#20 @arosen93
@utf: My apologies -- I didn't see the .html file that was actually generated. It worked perfectly despite the strange Ubuntu error message! I will admit, I have no clue what it's complaining about, but the image looks great!
