### 2021-03-12 fermisurfaces/IFermi#17 @arosen93
@utf: Thanks for the suggestion. I tried that and unfortunately received the same type of pyfftw issue.

```
   command: /home/asr731/software/anaconda/envs/ifermi/bin/python3.9 -u -c 'import sys, setuptools, tokenize; sys.argv[0] = '"'"'/tmp/pip-install-8093c5lm/pyfftw_978ce030d3f445cf84d51c59d935dbd8/setup.py'"'"'; __file__='"'"'/tmp/pip-install-8093c5lm/pyfftw_978ce030d3f445cf84d51c59d935dbd8/setup.py'"'"';f=getattr(tokenize, '"'"'open'"'"', open)(__file__);code=f.read().replace('"'"'\r\n'"'"', '"'"'\n'"'"');f.close();exec(compile(code, __file__, '"'"'exec'"'"'))' bdist_wheel -d /tmp/pip-wheel-60bfb9be
       cwd: /tmp/pip-install-8093c5lm/pyfftw_978ce030d3f445cf84d51c59d935dbd8/
  Complete output (30 lines):
  running bdist_wheel
  running build
  running build_py
  creating build
  creating build/lib.linux-x86_64-3.9
  creating build/lib.linux-x86_64-3.9/pyfftw
  copying pyfftw/__init__.py -> build/lib.linux-x86_64-3.9/pyfftw
  copying pyfftw/_version.py -> build/lib.linux-x86_64-3.9/pyfftw
  copying pyfftw/config.py -> build/lib.linux-x86_64-3.9/pyfftw
  creating build/lib.linux-x86_64-3.9/pyfftw/builders
  copying pyfftw/builders/__init__.py -> build/lib.linux-x86_64-3.9/pyfftw/builders
  copying pyfftw/builders/_utils.py -> build/lib.linux-x86_64-3.9/pyfftw/builders
  copying pyfftw/builders/builders.py -> build/lib.linux-x86_64-3.9/pyfftw/builders
  creating build/lib.linux-x86_64-3.9/pyfftw/interfaces
  copying pyfftw/interfaces/__init__.py -> build/lib.linux-x86_64-3.9/pyfftw/interfaces
  copying pyfftw/interfaces/_utils.py -> build/lib.linux-x86_64-3.9/pyfftw/interfaces
  copying pyfftw/interfaces/cache.py -> build/lib.linux-x86_64-3.9/pyfftw/interfaces
  copying pyfftw/interfaces/dask_fft.py -> build/lib.linux-x86_64-3.9/pyfftw/interfaces
  copying pyfftw/interfaces/numpy_fft.py -> build/lib.linux-x86_64-3.9/pyfftw/interfaces
  copying pyfftw/interfaces/scipy_fft.py -> build/lib.linux-x86_64-3.9/pyfftw/interfaces
  copying pyfftw/interfaces/scipy_fftpack.py -> build/lib.linux-x86_64-3.9/pyfftw/interfaces
  UPDATING build/lib.linux-x86_64-3.9/pyfftw/_version.py
  set build/lib.linux-x86_64-3.9/pyfftw/_version.py to '0.12.0'
  running build_ext
  Build pyFFTW with support for FFTW with
  double precision + openMP
  single precision + openMP
  long precision + openMP
  building 'pyfftw.pyfftw' extension
  error: unknown file type '.pyx' (from 'pyfftw/pyfftw.pyx')
  ERROR: Failed building wheel for pyfftw
  Running setup.py clean for pyfftw
Failed to build pyfftw
```
