### 2021-03-12 fermisurfaces/IFermi#17 @arosen93
Using `pip install cython` resolves the above issue. See https://github.com/pyFFTW/pyFFTW/issues/252.
