### 2021-03-12 fermisurfaces/IFermi#17 @utf
Ok, seems like:
```
conda create --name ifermi pip cmake numpy
conda activate ifermi
conda install -c conda-forge pymatgen boltztrap2 pyfftw
pip install ifermi
```
Should do the trick if people are having problems.
