### 2021-03-12 fermisurfaces/IFermi#17 @arosen93
Can confirm. Just tried that (after clearing out my prior environment), and it worked like a charm.
