### 2021-03-13 fermisurfaces/IFermi#21 @arosen93
JOSS Review: Optional flag to plot (but not show) a Fermi surface
RE: openjournals/joss-reviews#3089

**Describe the bug**
Upon running `ifermi plot`, it appears that an .html file is generated and an attempt is made to view the plot. I think it would be nice to have an optional flag to generate the plot but not show it. In my case, I'm SSH'd into a Linux machine without x-forwarding, and the code calls a web browser that I can't actually view and have to manually cancel out of every time I generate a plot via the command line interface.
