### 2021-03-13 fermisurfaces/IFermi#21 @utf
Also you can use `--output filename.jpg` to save a file rather generating an interactive plot.
