### 2021-03-13 fermisurfaces/IFermi#21 @arosen93
The second case would be really useful to add, I think! You can't beat the interactiveness :)
