### 2021-03-13 0ff4c3f
Enable saving html plots with plotly
