+++ ifermi/__init__.py
@@ -1,3 +1,3 @@
 """IFermi is a package for the generation, analysis and plotting of Fermi surfaces."""
 
-__version__ = "0.2.1"
+__version__ = "0.2.2"
