### 2021-03-13 fermisurfaces/IFermi#22 @arosen93
JOSS Review: --reciprocal-cell command line argument not supported
RE: openjournals/joss-reviews#3089

**Describe the bug**
Running `ifermi plot --reciprocal-cell` (as noted [here](https://fermisurfaces.github.io/IFermi/cli.html#reciprocal-space)) yields `Error: no such option: --reciprocal-cell`. Indeed, `ifermi plot -h` indicates it is not present in the list (nor is `-r`). Is this supposed to be `--no-wigner`?
