### 2021-03-13 fermisurfaces/IFermi#22 @utf
Great thanks! This has now been fixed (the documentation may take a few minutes to update).
