### 2021-03-13 fermisurfaces/IFermi#21 @utf
Ok, this has been added. Once you have finished your review I will push a new version with this feature.
