### 2021-03-13 fermisurfaces/IFermi#23 @utf
Thanks @arosen93. This is a great catch. I will update the script to include the projections.
