### 2021-03-15 #3089 @arosen93
I have concluded my review of IFermi. In short, IFermi is an excellent software package that I believe should be published in JOSS.

Overall, I am incredibly pleased with the detailed documentation, easy-to-follow tutorials, and accessible code layout. I believe that this will become the de facto tool for generating and analyzing Fermi surfaces for VASP users (and hopefully users of other codes once these features are added). I raised a few minor issues on the GitHub page, mainly with regards to the documentation, that have all been addressed to my satisfaction.

With regards to the accompanying paper, it clearly explains the purpose of IFermi. For comparison's sake, I tried out what I feel to be one of the closest packages to IFermi, PyProcar, to compare the two codes. Like IFermi, PyProcar has its own features for visualizing Fermi surface and allows for surface coloring based on properties. While there are many similarities, I came away from this exercise feeling that IFermi is still an important contribution to the field. As the authors aptly note in the accompanying manuscript, it is quite easy to analyze and modify the Fermi surfaces once generated with IFermi (since it's stored as a very convenient Python object). This was a clear advantage to me. The claim of novelty regarding the dimensionality determination is also accurate to the best of my knowledge. All in all, I agree with the statement of need and advertised features of IFermi.

Thank you for the opportunity to review IFermi. I am happy to recommend its acceptance in JOSS.
