### 2021-03-16 #3089 @lucydot
@danielskatz @utf

I finished reviewing iFermi today. I ran through the tutorial using my own dataset, so that I could compare it to the corresponding bandstructure that I have generated elsewhere. The iFermi output is consistent with what I'd expect looking at my 2D slice through reciprocal space.

To re-iterate what Andrew has already said, iFermi has easy to follow, comprehensive documentation. The CLI is intuitive to use, and installation is straight forward. There are similarities with existing packages, but these are highlighted in the JOSS paper, and the flexibility of iFermi distinguishes itself from what is already out there.

Last week I raised some minor issues regarding documentation, testing and installation. All of these have been addressed are now closed.

I'm happy to recommend iFermi for publication with JOSS. Thanks @utf @ajsearle97 for the code I'm looking forward to using it for future research
