### 2021-03-16 #3089 @utf
Hi @danielskatz

The archived version is v0.2.3
The zenodo doi is [10.5281/zenodo.4609270](https://zenodo.org/record/4609271)

I can confirm that I have checked the zenodo metadata.

Thanks you @lucydot and @arosen93 for the swift reviews. The package is definitely improved after your comments.
