### 2021-03-17 #3089 @utf
Hi @danielskatz, the paper pdf isn't loading on https://joss.theoj.org/papers/10.21105/joss.03089

Just checking if this is usual and I need to wait a little bit?
