### 2021-03-17 #3089 @arosen93
Very odd! Well, here's a sneak peak I guess! Works fine for me on Chrome and Edge even though you get the 404. Hopefully @danielskatz can provide some insight.

![image](https://user-images.githubusercontent.com/8674072/111402743-c2b9b200-8699-11eb-9244-d096b78b81ee.png)
