### 2021-03-17 #3089 @danielskatz
It worked for me on Safari - a check of it is the last thing I do before closing an issue.  It's likely a caching problem somewhere - If you are on a wifi network, you could try on a phone network or something like that, or just wait a bit.
